Final-Year-Projects/2026/MV-GSB_Vipin_Aryan/3_Code/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
scikit-image>=0.22.0
lpips>=0.1.4
